IT102-KS25_CNTT6_bt3.c

+++ IT102-KS25_CNTT6_bt3.c
@@ -6,9 +6,9 @@ int main() {
     int tong = a + b;
     int hieu = a - b;
     int tich = a * b;
-    int thuong = a / b;
+    float thuong = a / b;
     printf("tong = %d\n", tong);
     printf("hieu = %d\n", hieu);
     printf("tich = %d\n", tich);
-    printf("thuong = %d\n", thuong);
+    printf("thuong = %.3f\n", thuong);
 }
